+++ src/downstream.rs
@@ -4,9 +4,10 @@ use std::net::SocketAddr;
 use crate::body::Body;
 use crate::error::DownstreamRequestError;
 use crate::handoff::HandoffInfo;
+use crate::sandbox::PendingResponse;
 use http::{HeaderMap, Request, Response};
 use hyper::Uri;
-use tokio::sync::oneshot::Sender;
+use tokio::sync::mpsc::{self, Receiver, Sender};
 
 #[derive(Debug)]
 pub struct DownstreamMetadata {
@@ -33,10 +34,38 @@ pub struct DownstreamRequest {
     pub sender: Sender<DownstreamResponse>,
 }
 
+impl DownstreamRequest {
+    pub fn new(
+        req: hyper::Request<Body>,
+        metadata: DownstreamMetadata,
+    ) -> (Self, Receiver<DownstreamResponse>) {
+        let (sender, receiver) = mpsc::channel(10);
+        let req = Self {
+            req,
+            metadata,
+            sender,
+        };
+
+        (req, receiver)
+    }
+}
+
 #[derive(Debug)]
 pub enum DownstreamResponse {
+    /// An HTTP response to send back downstream.
     Http(Response<Body>),
+
+    /// A potential [Response] to an in-flight request that should be sent
+    /// downstream once it resolves.
+    ///
+    /// If the in-flight request fails, then the `Error` will be converted
+    /// to an appropriate 5XX HTTP status code.
+    Pending(Box<PendingResponse>),
+
+    /// Redirect the WebSocket request to pushpin.
     HandoffToPushpin(HandoffInfo),
+
+    /// Redirect the WebSocket request to a backend.
     HandoffToBackend(HandoffInfo),
 }