+++ src/wiggle_abi/resp_impl.rs
@@ -7,10 +7,10 @@ use {
         upstream,
         wiggle_abi::{
             fastly_http_resp::FastlyHttpResp,
-            headers::HttpHeaders,
+            headers::{HttpHeadersRead, HttpHeadersWrite},
             types::{
                 BodyHandle, FramingHeadersMode, HttpKeepaliveMode, HttpStatus, HttpVersion,
-                MultiValueCursor, MultiValueCursorResult, ResponseHandle,
+                MultiValueCursor, MultiValueCursorResult, PendingRequestHandle, ResponseHandle,
             },
         },
     },
@@ -104,7 +104,7 @@ impl FastlyHttpResp for Sandbox {
         value: GuestPtr<[u8]>,
     ) -> Result<(), Error> {
         let headers = &mut self.response_parts_mut(resp_handle)?.headers;
-        HttpHeaders::insert(headers, memory, name, value)
+        HttpHeadersWrite::insert(headers, memory, name, value)
     }
 
     fn header_append<'a>(
@@ -115,7 +115,7 @@ impl FastlyHttpResp for Sandbox {
         value: GuestPtr<[u8]>,
     ) -> Result<(), Error> {
         let headers = &mut self.response_parts_mut(resp_handle)?.headers;
-        HttpHeaders::append(headers, memory, name, value)
+        HttpHeadersWrite::append(headers, memory, name, value)
     }
 
     fn header_remove<'a>(
@@ -125,7 +125,7 @@ impl FastlyHttpResp for Sandbox {
         name: GuestPtr<[u8]>,
     ) -> Result<(), Error> {
         let headers = &mut self.response_parts_mut(resp_handle)?.headers;
-        HttpHeaders::remove(headers, memory, name)
+        HttpHeadersWrite::remove(headers, memory, name)
     }
 
     fn version_get(
@@ -150,7 +150,7 @@ impl FastlyHttpResp for Sandbox {
         Ok(())
     }
 
-    fn send_downstream(
+    async fn send_downstream(
         &mut self,
         _memory: &mut GuestMemory<'_>,
         resp_handle: ResponseHandle,
@@ -168,7 +168,16 @@ impl FastlyHttpResp for Sandbox {
             };
             Response::from_parts(resp_parts, body)
         }; // Set the downstream response, and return.
-        self.send_downstream_response(resp)
+        self.send_downstream_response(resp).await
+    }
+
+    async fn send_downstream_pending(
+        &mut self,
+        _memory: &mut GuestMemory<'_>,
+        h: PendingRequestHandle,
+    ) -> Result<(), Error> {
+        let pending = self.take_pending_request(h)?;
+        self.send_pending_response(pending).await
     }
 
     fn status_get(